refactor: コンポーネント以下の配置の変更
