--- src/components/TempComponent.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-import { Card, CardBody, Text } from "@chakra-ui/react";
-
-export const TempComponent = () => {
-  return (
-    <Card>
-      <CardBody>
-        <Text>Hello Next.js</Text>
-      </CardBody>
-    </Card>
-  );
-};
